addresses https://github.com/Renascence-Labs/2024-01-locksmith/issues/7
Instead of taking the additional gas cost of looping through the ids and
reverting on duplicates, I decided to call super._update FIRST, so all
the tokens could be moved. In this way, all I have to do is test the
resulting balance and trip if the amount is breached.

This enables more composability because sending two keys via batch
should be possible, and we don't want to diverge from the interface
expectations if we don't have to. Also, not looping through the key ids
n^2 complexity saves more gas on this operation.

I've updated _manageIndexes() as well to reflect this new reality,
removing issues about the holder index at the same time.